feat: add "module" field to package.json
This commit adds the "module" field to the package.json file. The "module" field specifies the location of the ES module version of the library, which is used by modern JavaScript build tools like webpack to enable tree-shaking and other optimizations.

The "module" field points to the index.js file in the "src" directory, which exports the same public API as the main index.js file in the "dist" directory, but in ES module format.

This change will make it easier for developers to use this library in modern JavaScript projects, and improve the performance and bundle size of their applications.