chore: exclude the dist folder
Because it is a generated folder that can be easily recreated during the build process. The purpose of the dist folder is to contain the compiled code that can be used in production, but it is not needed for development.

Including the dist folder in version control can cause several issues:

It increases the size of the repository unnecessarily, making it slower to clone and work with.

It can cause merge conflicts when different developers are working on the same codebase and generate different versions of the dist folder.

It can lead to versioning issues when trying to release new versions of the library, as the compiled code in the dist folder may not match the source code.